refactor(video): Let a texture size say it's never negative
The SDL texture held its size in signed members behind accessors the base class
declares unsigned and both renderers kept SDL_CreateWindow's flags in an int.
Correcting those types and crossing over once where SDL hands us a surface
accounts for most of this as neither a texture nor a window flag is ever negative.

The rest are indexes and counters reaching a container. A pixel height that had
been declared a string's size_type, a vertex counter, and the string offsets in
the two wrapping helpers. Sprite::set_frame goes rather than gains a cast, since
nothing calls it and its own modulo would have turned a negative frame into a
very large one.

+++ src/video/sdl/sdl_painter.cpp
@@ -287,7 +287,7 @@ SDLPainter::draw_filled_rect(SDL_Renderer* renderer, const DrawingRequest& reque
 
     // rounded top and bottom parts
     std::vector<SDL_Rect> rects;
-    rects.reserve(2*slices + 1);
+    rects.reserve(static_cast<size_t>(2*slices + 1));
     for(int i = 0; i < slices; ++i)
     {
       float p = (static_cast<float>(i) + 0.5f) / static_cast<float>(slices);
